ECCI-494: Fix memory error caused by a node being its own parent
This amends the `isAncestorRestricted` method of `localgov_restricted_content` to detect cases where a node has been set as its direct parent, and breaks the parent-recursion chain by restricting the content. It won't detect cases where there's a cycle of length > 1, but this allows editing of a page on the intranet where this was already set.